requirements.txt
pygame
langchain
openai
python-dotenv
